refactor: scope_reference.rs から ui-sref / ng-model 処理を別モジュールへ分離 (#48)
`scope_reference.rs` がファイル名の責務 (\$scope 参照収集) を超えて以下も
処理していたため、責務ごとにモジュールを分離する:

- ui-router の `ui-sref` 属性値からの state 名参照登録
- `ng-model` の暗黙的な `\$scope` 定義登録 (HtmlNgModelTarget)

ファイル名と実装内容が乖離していたため、ファイルを開いた瞬間に何を扱っているかが
分からない状態だった。既存の `directive_reference.rs` / `form.rs` のように責務
単位で分けることで、新しい属性 (例: `track-by`) 対応を追加するときの場所も明確
になる。

変更:
- `src/analyzer/html/ui_sref.rs` を新規追加
  - `register_ui_sref_reference` を移動 (`scope_reference.rs` から)
- `src/analyzer/html/ng_model.rs` を新規追加
  - `register_ng_model_target` を新規追加 (旧コードはインライン展開されていた)
- `scope_reference.rs` 側は両ヘルパーを呼ぶだけにし、import から
  `HtmlNgModelTarget` / `HtmlUiSrefReference` / `Span` / `SymbolReference` を削除
- `mod.rs` に新モジュール 2 つを登録

挙動変更なし (移動のみ)。テスト: lib 161 + integration 148 + investigate 2
すべて通過。

Closes #48
